Fix for CursorTo method
The original code has swapped the roles of x and y, and was working
fine if x == y or if the roles of x and y where swapped.
Each line having a specific offset, the values are simple gathered in
an array indexed by the value of y. Invalid values move the cursor to
home.